# Dokumenti/OpenAPISpec/niapi-openapi3.yaml
openapi: 3.0.1
info:
  version: 0.0.1
  title: niapi
  description: Specifikacija APIja koje implementira API servis narodnih inicijativa Srbije
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/elektronske-narodne-inicijative/niapi/0.0.1
  - url: https://einicijative.gov.rs/niapi/v1
paths:
...
  /ovlice/profil:
    get:
      description: Upit detalja profila ovlašćenog lica
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Detalji profila ovlašćenog lica
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OvliceProfilOdgovor'
        '400':
          $ref: '#/components/responses/400Error'
        '401':
          $ref: '#/components/responses/401Error'
        '500':
          $ref: '#/components/responses/500Error'
  /ovlice/inicijative/zaodobrenje:
    get:
      description: Lista inicijativa koje čekaju odobrenje na odgovarajućem nivou 
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista inicijativa koje čekaju odobrenje
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OvliceListaZaOdobrenjeOdgovor'
        '400':
          $ref: '#/components/responses/400Error'
        '401':
          $ref: '#/components/responses/401Error'
        '500':
          $ref: '#/components/responses/500Error'
  /ovlice/inicijative/zaishod:
    get:
      description: Lista inicijativa koje čekaju ishod rasprave u skupštini na odgovarajućem nivou 
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista inicijativa koje čekaju ishod
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OvliceListaZaIshodOdgovor'
        '400':
          $ref: '#/components/responses/400Error'
        '401':
          $ref: '#/components/responses/401Error'
        '500':
          $ref: '#/components/responses/500Error'
  /ovlice/inicijativa/{idInicijative}:
    get:
      description: Detalji inicijative
      security:
        - bearerAuth: []
      parameters:
        - name: idInicijative
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Detalji inicijative
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OvliceInicijativaOdgovor'
        '400':
          $ref: '#/components/responses/400Error'
        '401':
          $ref: '#/components/responses/401Error'
        '500':
          $ref: '#/components/responses/500Error'
  /ovlice/inicijativa/{idInicijative}/odbij:
    put:
      description: Ovlašćeno lice odbija zahtev za narodnu inicijativu uz obrazloženje
      security:
        - bearerAuth: []
      parameters:
        - name: idInicijative
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Obrazlozenje'
      responses:
        '200':
          description: Uspeh
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Uspeh'
        '400':
          $ref: '#/components/responses/400Error'
        '401':
          $ref: '#/components/responses/401Error'
        '500':
          $ref: '#/components/responses/500Error'
  /ovlice/inicijativa/{idInicijative}/odobri:
    put:
      description: Ovlašćeno lice odobrava zahtev za narodnu inicijativu uz opciono obrazloženje
      security:
        - bearerAuth: []
      parameters:
        - name: idInicijative
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Uspeh
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Uspeh'
        '400':
          $ref: '#/components/responses/400Error'
        '401':
          $ref: '#/components/responses/401Error'
        '500':
          $ref: '#/components/responses/500Error'
  /ovlice/inicijativa/{idInicijative}/prihvacena:
    put:
      description: Ovlašćeno lice obeležava da je pokrenuta narodna inicijativa prihvaćena, uz opciono obrazloženje (beleške sa sednice skupštine)
      security:
        - bearerAuth: []
      parameters:
        - name: idInicijative
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ObrazlozenjeSaSednice'
      responses:
        '200':
          description: Uspeh
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Uspeh'
        '400':
          $ref: '#/components/responses/400Error'
        '401':
          $ref: '#/components/responses/401Error'
        '500':
          $ref: '#/components/responses/500Error'
  /ovlice/inicijativa/{idInicijative}/odbacena:
    put:
      description: Ovlašćeno lice obeležava da je pokrenuta narodna inicijativa odbačena od strane skupštine, uz opciono obrazloženje (beleške sa sednice skupštine)
      security:
        - bearerAuth: []
      parameters:
        - name: idInicijative
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ObrazlozenjeSaSednice'
      responses:
        '200':
          description: Uspeh
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Uspeh'
        '400':
          $ref: '#/components/responses/400Error'
        '401':
          $ref: '#/components/responses/401Error'
        '500':
          $ref: '#/components/responses/500Error'

  #  ----------------------------------------------------------
  # Metodi namenjeni članovima inicijativnih odbora, javni API
  #  ----------------------------------------------------------
...
components:
  securitySchemes:

    basicAuth:
      type: http
      scheme: basic

    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  #  -------------------------------------------------------
  #  Definicije struktura podataka
  #  -------------------------------------------------------
  schemas:
...
    Uspeh:
      type: object
      properties:
        poruka:
          type: string
    Obrazlozenje:
      type: object
      properties:
        obrazlozenje:
          type: string
    ObrazlozenjeSaSednice:
      type: object
      required:
        - datumSednice
      properties:
        datumSednice:
          type: string
          format: date
        obrazlozenje:
          type: string
...
    OsnovnoOInicijativi: # za tabelarni prikaz
      type: object
      required:
        - idInicijative
        - nazivInicijative
        - tipInicijative
        - trnZahteva
        - brInicijatora
      properties:
        idInicijative:
          type: integer
        nazivInicijative:
          type: string
        tipInicijative:
          type: string
        trnZahteva:
          type: string
          format: date-time
        trnPodnosenja:
          type: string
          format: date-time
        brInicijatora:
          type: integer
        datumPokretanja:
          type: string
          format: date
        brPotpisa:
          type: integer
    InicijativaDetalji:  # za detaljni prikaz
      type: object
      required:
        - idInicijative
        - tipInicijative
        - inicijator
        - nazivInicijative
        - tekstInicijative
        - trnZahteva
        - fazaObrade
        - emailZaKontakt
        - nivoVlasti
      properties:
        idInicijative:
          type: integer
        tipInicijative:
          type: string
        inicijator:
          type: string
        nazivInicijative:
          type: string
        tekstInicijative:
          type: string
        fazaObrade:
          type: string
        emailZaKontakt:
          type: string
        nivoVlasti:
          type: string
        jedinicaVlasti:
          type: string
        trnZahteva:
          type: string
          format: date-time
        trnOdbijanjaZahteva:
          type: string
          format: date-time
        razlogOdbijanjaZahteva:
          type: string
        trnPodnosenja:
          type: string
          format: date-time
        datumPokretanja:
          type: string
          format: date
        datumOdluke:
          type: string
          format: date
        prihvacena:
          type: boolean
        beleskaSaSednice:
          type: string
        clanoviInicijativnogOdbora:
          type: array
          items:
            $ref: '#/components/schemas/ClanInicijativnogOdbora'
        prilozi:
          type: array
          items:
            $ref: '#/components/schemas/PrilogInicijative'
        promene:
          type: array
          items:
            $ref: '#/components/schemas/DnevnikPromena'
    ClanInicijativnogOdbora:
      type: object
      required:
        - aktivan
        - imePrezime
        - pol
        - godinaRodjenja
      properties:
        aktivan:
          type: boolean
        imePrezime:
          type: string
        pol:
          type: string
        godinaRodjenja:
          type: integer
        emailZaKontakt:
          type: string
        biografija:
          type: string
        trnPrihvatanjaClanstva:
          type: string
          format: date-time
        trnOdustajanjaOdClanstva:
          type: string
          format: date-time
    PrilogInicijative:
      type: object
      required:
        - nazivPriloga
        - urlPriloga
        - sortiranje
        - obrisan
      properties:
        nazivPriloga:
          type: string
        urlPriloga:
          type: string
        sortiranje:
          type: integer
        obrisan:
          type: boolean
    DnevnikPromena:
      type: object
      required:
        - trnPromene
        - fazaObrade
        - detaljiPromene
      properties:
        trnPromene:
          type: string
          format: date-time
        fazaObrade:
          type: string
        detaljiPromene:
          type: string

    #
    #  Strukture za salterske transakcije
    #
...
    OvliceProfilOdgovor:
      type: object
      required:
        - imePrezime
        - emailAdresa
        - nivoUprave
      properties:
        imePrezime:
          type: string
        emailAdresa:
          type: string
        nivoUprave:
          type: string
        opisJediniceUprave:
          type: string
    OvliceListaZaOdobrenjeOdgovor:
      type: object
      properties:
        inicijativeZaOdobrenje:
          type: array
          items:
            $ref: '#/components/schemas/OsnovnoOInicijativi'
    OvliceListaZaIshodOdgovor:
      type: object
      properties:
        inicijativeZaIshod:
          type: array
          items:
            $ref: '#/components/schemas/OsnovnoOInicijativi'
    OvliceInicijativaOdgovor:
      type: object
      properties:
        detaljiInicijative:
            $ref: '#/components/schemas/InicijativaDetalji'

    #
    #  Strukture za transakcije inicijatora i članova inicijativnih odbora
    #
...
  responses:
    400Error:
      description: Aplikativne greske (pogresan sadrzaj)
      content:
        application/json:
          schema:
            type: object
            properties:
              resultType:
                type: string
                enum:
                  - PogresanZahtev
                  - PogresniPodaciUZahtevu
                  - NepoznatPotpisnik
                  - NepoznatInicijator
                  - NepoznatoOvlascenoLice
              resultCode:
                type: string
    401Error:
      description: Neispravna ili istekla sesija
      content:
        application/json:
          schema:
            type: object
            properties:
              resultType:
                type: string
                enum:
                  - NeispravnaSesija
                  - IsteklaSesija
              resultCode:
                type: string
    500Error:
      description: Neocekivane programske greske
      content:
        application/json:
          schema:
            type: object
            properties:
              resultType:
                type: string
                enum:
                  - ProblemUObradi
              resultCode:
                type: string